app/requirements.txt
google-adk
python-dotenv
vertexai
